backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.4
cron-descriptor==1.4.5
cssselect2==0.9.0
dj-database-url==3.1.2
Django==6.0.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==40.11.0
fonttools==4.62.1
gunicorn==25.1.0
inflection==0.5.1
iniconfig==2.3.0
iso3166==2.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==6.0.2
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.12.1
pyphen==0.17.2
pytest==9.0.2
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-stdnum==2.2
PyYAML==6.0.3
qrbill==1.2.0
qrcode==8.2
redis==7.3.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
svgwrite==1.4.3
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
# MariaDB/MySQL support: install libmariadb-dev then add mysqlclient
